backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
pyyaml>=6.0
aiofiles>=24.1.0
chromadb>=0.5.0
pypdf>=4.3.0
python-docx>=1.1.0
markdown>=3.6
beautifulsoup4>=4.12.0
httpx>=0.27.0
websockets>=12.0
